[SPARK-23425][SQL]When wild card is been used  in load command system is throwing  analysis exception
## What changes were proposed in this pull request?
Check if the path exists or there are matched paths if it's a path with wildcard.For HDFS path, we support wildcard in directory name and file name.
fs.exists(srcPath) API cannot resolve the path with wild card pattern, so globStatus() api is been used which can resolve the paths with hdfs supported wildcard
string like *,? etc.

## How was this patch tested?
Manually tested in hdfs-Yarn clustaer
